feat: added base for admin view of records page
